Create multi_cte
This code was used to create output that shows the top 5 customers in rental revenue from the top10 cities in rental revenue from the top 5 countries in total revenue. Two of the top 5 customers were from different cities in Mexico. All of the data is dummy data; no PII for actual people was used.